+++ common/models/src/main/java/app/ss/models/VideoAux.kt
@@ -20,17 +20,32 @@
  * THE SOFTWARE.
  */
 
-package io.adventech.blockkit.model
+package app.ss.models
 
-import androidx.annotation.Keep
 import com.squareup.moshi.JsonClass
 
-@Keep
 @JsonClass(generateAdapter = true)
-data class PDFAux(
+data class VideoAuxArtist(
     val id: String,
-    val src: String,
+    val artist: String,
     val title: String,
-    val target: String?,
-    val targetIndex: String?
+    val target: String,
+    val targetIndex: String,
+    val src: String,
+    val thumbnail: String,
+)
+
+@JsonClass(generateAdapter = true)
+data class VideoAux(
+    val artist: String,
+    val clips: List<VideoAuxArtist>,
+)
+
+@JsonClass(generateAdapter = true)
+data class VideoClipSegment(
+    val src: String,
+    val artist: String?,
+    val title: String?,
+    val thumbnail: String?,
+    val hls: String?,
 )
