+++ services/resources/api/src/main/kotlin/ss/resources/api/test/FakeResourcesRepository.kt
@@ -23,9 +23,9 @@
 package ss.resources.api.test
 
 import androidx.annotation.VisibleForTesting
-import io.adventech.blockkit.model.AudioAux
-import io.adventech.blockkit.model.PDFAux
-import io.adventech.blockkit.model.VideoAux
+import app.ss.models.AudioAux
+import app.ss.models.PDFAux
+import app.ss.models.VideoAux
 import io.adventech.blockkit.model.feed.FeedGroup
 import io.adventech.blockkit.model.feed.FeedType
 import io.adventech.blockkit.model.resource.Resource
