+++ services/resources/impl/src/main/kotlin/ss/resources/impl/ResourcesRepositoryImpl.kt
@@ -22,12 +22,12 @@
 
 package ss.resources.impl
 
+import app.ss.models.AudioAux
+import app.ss.models.PDFAux
+import app.ss.models.VideoAux
 import app.ss.network.NetworkResource
 import app.ss.network.safeApiCall
 import dagger.Lazy
-import io.adventech.blockkit.model.AudioAux
-import io.adventech.blockkit.model.PDFAux
-import io.adventech.blockkit.model.VideoAux
 import io.adventech.blockkit.model.feed.FeedGroup
 import io.adventech.blockkit.model.feed.FeedType
 import io.adventech.blockkit.model.resource.Resource
